Update README.md
Weather site: when loaded, it shows the weather for today. When you click on the link in the navbar "five days" - shows the weather for the next five days. When you click on one of the blocks of five days - shows the weather of this day with an interval of three hours. You can enter any city in the search bar, and when you click on the "search" button, the data will be updated. If you enter a non-existent city, then an error message will appear on the screen, which will notify the user about this, and will disappear automatically after three seconds.